# smartgas-payments-composer/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Smart Gas API Specification",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/elegibility/customers/levels": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Paginate Customer Levels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "Customer Levels List",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Lookup in customer levels",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer Levels Paginated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.CustomerLevelListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/elegibility/customers/levels/": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create customer level with given parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "Create customer level",
                "parameters": [
                    {
                        "description": "Customer Level create body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ElegibilityCreateCustomerLevelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CustomerLevelCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "406": {
                        "description": "Foreign key for level, customer not exists",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/elegibility/customers/levels/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update customer level with given parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "Update customer level",
                "parameters": [
                    {
                        "description": "Customer Level Update body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ElegibilityUpdateCustomerLevelRequest"
                        }
                    },
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "uuid4 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "406": {
                        "description": "Foreign key for level, customer not exists",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/elegibility/levels": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Paginate Levels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "List Levels",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Lookup in levels",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Levels Paginated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.LevelListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create Elegibility level with given parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "Create Elegibility Level",
                "parameters": [
                    {
                        "description": "Elegibility body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EelegibilityLevelCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.LevelCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/elegibility/levels/all": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List of all levels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "List all levels",
                "responses": {
                    "200": {
                        "description": "Levels",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LevelListAllResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/elegibility/levels/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Elegibility Level with given parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elegibility"
                ],
                "summary": "Update Elegibility Level",
                "parameters": [
                    {
                        "description": "Campaign body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ElegibilityLevelUpdatesRequest"
                        }
                    },
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "uuid4 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BadRequestMessage": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CustomerLevelCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.CustomerLevelListResponse": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "elegibility_level_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "validity_month": {
                    "type": "integer"
                },
                "validity_year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ElegibilityCreateCustomerLevelRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "elegibility_level_id",
                "validity_month",
                "validity_year"
            ],
            "properties": {
                "customer_id": {
                    "type": "string",
                    "example": "23ae8c18-4d7a-41a3-a148-8ae2d0a75690"
                },
                "elegibility_level_id": {
                    "type": "string",
                    "example": "23ae8c18-4d7a-41a3-a148-8ae2d0a75690"
                },
                "validity_month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "validity_year": {
                    "type": "integer",
                    "maximum": 2030,
                    "minimum": 2015
                }
            }
        },
...
        "dto.GeneralMessage": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "format": "array"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and paste the access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}